Moved simu LCD buffer to simu target
