+++ radio/src/gui/colorlcd/lcd.cpp
@@ -85,84 +85,6 @@ void lcdSetContrast()
   lcdSetRefVolt(g_eeGeneral.contrast);
 }
 
-#if defined(SIMU)
-BitmapBuffer _lcd(BMP_RGB565, LCD_W, LCD_H, displayBuf);
-BitmapBuffer * lcd = &_lcd;
-
-void DMAFillRect(uint16_t * dest, uint16_t destw, uint16_t desth, uint16_t x, uint16_t y, uint16_t w, uint16_t h, uint16_t color)
-{
-#if defined(LCD_VERTICAL_INVERT)
-  x = destw - (x + w);
-  y = desth - (y + h);
-#endif
-
-  for (int i=0; i<h; i++) {
-    for (int j=0; j<w; j++) {
-      dest[(y+i)*destw+x+j] = color;
-    }
-  }
-}
-
-void DMACopyBitmap(uint16_t * dest, uint16_t destw, uint16_t desth, uint16_t x, uint16_t y, const uint16_t * src, uint16_t srcw, uint16_t srch, uint16_t srcx, uint16_t srcy, uint16_t w, uint16_t h)
-{
-#if defined(LCD_VERTICAL_INVERT)
-  x = destw - (x + w);
-  y = desth - (y + h);
-  srcx = srcw - (srcx + w);
-  srcy = srch - (srcy + h);
-#endif
-
-  for (int i=0; i<h; i++) {
-    memcpy(dest+(y+i)*destw+x, src+(srcy+i)*srcw+srcx, 2*w);
-  }
-}
-
-void DMACopyAlphaBitmap(uint16_t * dest, uint16_t destw, uint16_t desth, uint16_t x, uint16_t y, const uint16_t * src, uint16_t srcw, uint16_t srch, uint16_t srcx, uint16_t srcy, uint16_t w, uint16_t h)
-{
-#if defined(LCD_VERTICAL_INVERT)
-  x = destw - (x + w);
-  y = desth - (y + h);
-  srcx = srcw - (srcx + w);
-  srcy = srch - (srcy + h);
-#endif
-
-  for (coord_t line=0; line<h; line++) {
-    uint16_t * p = dest + (y+line)*destw + x;
-    const uint16_t * q = src + (srcy+line)*srcw + srcx;
-    for (coord_t col=0; col<w; col++) {
-      uint8_t alpha = *q >> 12;
-      uint8_t red = ((((*q >> 8) & 0x0f) << 1) * alpha + (*p >> 11) * (0x0f-alpha)) / 0x0f;
-      uint8_t green = ((((*q >> 4) & 0x0f) << 2) * alpha + ((*p >> 5) & 0x3f) * (0x0f-alpha)) / 0x0f;
-      uint8_t blue = ((((*q >> 0) & 0x0f) << 1) * alpha + ((*p >> 0) & 0x1f) * (0x0f-alpha)) / 0x0f;
-      *p = (red << 11) + (green << 5) + (blue << 0);
-      p++; q++;
-    }
-  }
-}
-
-void DMABitmapConvert(uint16_t * dest, const uint8_t * src, uint16_t w, uint16_t h, uint32_t format)
-{
-  if (format == DMA2D_ARGB4444) {
-    for (int row = 0; row < h; ++row) {
-      for (int col = 0; col < w; ++col) {
-        *dest = ARGB(src[0], src[1], src[2], src[3]);
-        ++dest;
-        src += 4;
-      }
-    }
-  }
-  else {
-    for (int row = 0; row < h; ++row) {
-      for(int col = 0; col < w; ++col) {
-        *dest = RGB(src[1], src[2], src[3]);
-        ++dest;
-        src += 4;
-      }
-    }
-  }
-}
-#endif
-
 event_t getWindowEvent()
 {
   return getEvent(false);
