+++ radio/src/targets/simu/CMakeLists.txt
@@ -15,6 +15,7 @@ set(SIMU_SRC
   simueeprom.cpp
   simufatfs.cpp
   simudisk.cpp
+  simulcd.cpp
   )
 
 if(SIMU_DISKIO)
