+++ radio/src/targets/simu/simulcd.h
@@ -26,5 +26,6 @@
 extern int g_snapshot_idx;
 extern bool simuLcdRefresh;
 extern pixel_t simuLcdBuf[DISPLAY_BUFFER_SIZE];
+extern pixel_t displayBuf[DISPLAY_BUFFER_SIZE];
 
 #endif // _SIMULCD_H_
